Correct omission of "attack helicopter" gender identity.
No distinction is made between attack helicopters with coaxial rotors such as
the Ka-50 Black Shark vs the AH-64 Apache's tail rotor design, but this should
not be interpreted as the author's rejection of the importance of this
to attack helicopters.  Future work should allow for specifying these details
as well as armament and operational range which determines who an attack
helicopter engages with.